requirements.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.0

# External Services
supabase==2.3.0
realtime==1.0.6
pinecone>=5.0.0
huggingface-hub>=0.20.0
langchain-community>=0.0.20
langchain>=0.1.0
pymupdf>=1.23.0
pypdf>=4.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0

